# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UHC Agentic Claims-Audit Backend
  description: |
    REST API for the Wipro UHC agentic claims-audit platform.

    Top-level groups:
      * `/api/builder/`     — workflow CRUD, server-driven palette + chrome, atomic canvas save, per-node attachments.
      * `/api/ingest/`      — SOP ingestion jobs, knowledge-graph + section data, narrative backfill, HTML viewer.
      * `/api/agent-tools/` — DB-backed LangChain tool registry + per-tool invoke surface (see AGENT_TOOLS.md).
      * `/api/execute/`     — Per-Shape execution engine: upload an Excel of claim ids and get a per-claim adjudication.

    Most endpoints require a `Bearer` JWT minted by the Node `claims-corebackend` (HS256, shared `JWT_SECRET`). Endpoints that allow anonymous access are flagged below.
  version: "1.1.0"
  contact:
    name: UHC Agentic Backend team
servers:
  - url: http://localhost:8000
    description: Local dev
  - url: https://uhc-backend.example.com
    description: Production placeholder
...
security:
  - bearerAuth: []
paths:
...
  /api/execute/batches/{batch_id}/events/:
    get:
      tags: [execute]
      summary: Server-Sent Events stream of batch progress
      description: |
        Subscribes to the Redis pub/sub channel `batch:<batch_id>` and pipes
        per-Shape (`shape_start`), per-rule (`rule_evaluated`), and per-claim
        (`claim`) events to the SPA as they happen. On connect, replays any
        already-finished claims from the DB so a reconnect picks up cleanly.

        Terminates when a `summary` (happy path) or `error` (task crash)
        event arrives. Each event is framed in SSE format
        (`event: <kind>\ndata: <json>\n\n`). The bridge sends a SSE comment
        `: keepalive` every 15 seconds when the channel is idle, to defeat
        proxy idle timeouts.

        OpenAPI cannot fully model an SSE stream, so the schema below is
        an approximation. See [docs/EXECUTION_ENGINE.md §5.6](EXECUTION_ENGINE.md)
        for the exact wire grammar.
      parameters:
        - name: batch_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: |
            SSE stream. Each chunk is one SSE-framed event whose `event:`
            line names the event kind. See `BatchEventsStream` for the
            possible event payloads (one of `batch_start`, `shape_start`,
            `rule_evaluated`, `claim`, `summary`, `error`).
          content:
            text/event-stream:
              schema: { $ref: "#/components/schemas/BatchEventsStream" }
        "404":
          description: |
            Unknown `batch_id`. Delivered as a single `event: error` SSE
            payload (status 404), since SSE responses can't easily 404
            mid-stream.
          content:
            text/event-stream:
              schema: { $ref: "#/components/schemas/BatchEventsStream" }

# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 JWT issued by the Node `claims-corebackend`.
...
  schemas:
...
    SseBatchStart:
      type: object
      properties:
        batch_id:        { type: string, format: uuid }
        workflow_id:     { type: string, format: uuid }
        total_claims:    { type: integer }
        claim_id_column: { type: string }
        source_filename: { type: string }
        status:          { type: string, enum: [RUNNING, COMPLETED, PARTIAL, FAILED] }
    SseShapeStart:
      type: object
      properties:
        batch_id:    { type: string, format: uuid }
        run_id:      { type: string, format: uuid }
        claim_id:    { type: string }
        shape_id:    { type: string }
        shape_label: { type: string }
        rules_total: { type: integer }
        ts:          { type: number, description: "Unix epoch seconds, float." }
    SseRuleEvaluated:
      type: object
      properties:
        batch_id:      { type: string, format: uuid }
        run_id:        { type: string, format: uuid }
        claim_id:      { type: string }
        shape_id:      { type: string }
        shape_label:   { type: string }
        rule_key:      { type: string }
        rule_source:   { type: string, enum: [PRECONDITION, DECISION] }
        matched:       { type: boolean }
        decision_type: { type: string }
        confidence:    { type: number, format: float }
        reasoning:     { type: string }
        codes:
          type: array
          items: { type: string }
        llm_provider:  { type: string }
        llm_model:     { type: string }
        llm_ms:        { type: integer }
        llm_attempts:
          type: integer
          description: How many physical LLM attempts the helper made for this rule (retries + cross-provider fallback collapse into one event).
        ts:            { type: number }
    SseClaim:
      type: object
      properties:
        claim_id:            { type: string }
        run_id:              { type: string, format: uuid }
        status:
          type: string
          enum: [RUNNING, COMPLETED, FAILED, TERMINATED_EARLY, FETCH_FAILED]
        final_decision_type: { type: string }
        applied_codes:
          type: array
          items: { type: string }
        narrative:     { type: string }
        error_message: { type: string }
    SseSummary:
      type: object
      properties:
        id:            { type: string, format: uuid }
        status:        { type: string, enum: [COMPLETED, PARTIAL, FAILED] }
        total_claims:  { type: integer }
        completed:     { type: integer }
        failed:        { type: integer }
        duration_ms:   { type: integer }
        error_message: { type: string }
    SseError:
      type: object
      properties:
        batch_id: { type: string, format: uuid }
        message:  { type: string }
    BatchEventsStream:
      description: |
        Approximation of the SSE event stream. Each chunk in the response
        is one SSE event of the form

            event: <kind>
            data: <json>

        followed by a blank line. `<kind>` is one of `batch_start`,
        `shape_start`, `rule_evaluated`, `claim`, `summary`, or `error`,
        with the payload shapes referenced below.
      oneOf:
        - $ref: "#/components/schemas/SseBatchStart"
        - $ref: "#/components/schemas/SseShapeStart"
        - $ref: "#/components/schemas/SseRuleEvaluated"
        - $ref: "#/components/schemas/SseClaim"
        - $ref: "#/components/schemas/SseSummary"
        - $ref: "#/components/schemas/SseError"